Maintenance
Clean up Windows build

+++ build_sdk.sh
@@ -126,10 +126,17 @@ cd "$path/android"
 	cp -v libs/armeabi-v7a/* "$OUTPUT_DIR_ANDROID"
 cd -
 
+if [ ! -f win32/Release/plugin_openssl.dll ]
+then
+	echo "Build plugin for win32 before packaging"
+	exit 1
+fi
+
 echo "------------------------------------------------------------------------"
 echo "[win32]"
 cd "$path"
 	cp -v sdk-openssl/win32/bin/*.dll "$OUTPUT_DIR_WIN32"
+	cp -v win32/Release/plugin_openssl.dll "$OUTPUT_DIR_WIN32"
 cd -
 
 if [ -d "$path/docs" ]
